requirements.txt
fastapi==0.127.1
uvicorn==0.40.0
sqlalchemy==2.0.45
asyncpg==0.31.0
greenlet==2.0.2
google-api-python-client==2.187.0
google-auth==2.45.0
jinja2==3.1.6
python-dotenv==1.2.1
pydantic==2.12.5
pydantic-settings==2.12.0
requests==2.32.5
